fix(auth): fix the type issue
