--- packages/auth/types.ts
@@ -1,25 +1,0 @@
-import type { DefaultSession } from "@auth/core/types";
-
-declare module "@auth/core/types" {
-  interface Session extends DefaultSession {
-    user: {
-      role: "admin" | "user";
-    } & DefaultSession["user"];
-  }
-
-  interface User {
-    role: "admin" | "user";
-  }
-}
-
-declare module "next-auth" {
-  interface Session extends DefaultSession {
-    user: {
-      role: "admin" | "user";
-    } & DefaultSession["user"];
-  }
-
-  interface User {
-    role: "admin" | "user";
-  }
-}
